Implementando lógica de sincronismo baseado no ping no dos eventos de SYNC;

+++ GameServerLib/Sources/GameNetworking/Public API/Server/GameServer.cs
@@ -51,7 +51,7 @@ namespace GameNetworking {
             });
         }
 
-        public int GetPing(NetworkPlayer player) {
+        public float GetPing(NetworkPlayer player) {
             return this.pingController.GetPingValue(player);
         }
